requirements.txt
streamlit
pandas
mathplotlib
utilidades
plotly.express
PIL
plotly.graph_objects